created calc interval as its own component
